requirements.txt
aerich==0.9.2
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncclick==8.3.0.3
click==8.3.0
dictdiffer==0.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.2
h11==0.16.0
idna==3.11
iso8601==2.1.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pypika-tortoise==0.6.2
python-dotenv==1.2.1
pytz==2025.2
sniffio==1.3.1
starlette==0.49.1
tortoise-orm==0.25.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
